rename IT.java Tests.java
