+++ community/memories/spring-ai-alibaba-starter-memory-jdbc/src/test/java/com/alibaba/cloud/ai/memory/jdbc/PostgresChatMemoryRepositoryTests.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package mongodb;
+package com.alibaba.cloud.ai.memory.jdbc;
 
-import com.alibaba.cloud.ai.memory.mongodb.MongoDBChatMemoryRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -23,49 +22,47 @@ import org.springframework.ai.chat.memory.ChatMemoryRepository;
 import org.springframework.ai.chat.messages.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.boot.autoconfigure.jdbc.JdbcTemplateAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Bean;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.MongoDBContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 import java.util.List;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-/**
- * Integration test using Testcontainers to automatically manage MongoDB test environment
- */
-@SpringBootTest(classes = MongoDBChatMemoryRepositoryIT.TestConfiguration.class)
+@SpringBootTest(classes = PostgresChatMemoryRepositoryTests.TestConfiguration.class)
 @Testcontainers
-class MongoDBChatMemoryRepositoryIT {
+class PostgresChatMemoryRepositoryTests {
 
-	private static final int MongoDB_PORT = 27017;
-
-	// Define and start MongoDB container
 	@Container
-	private static final MongoDBContainer mongoDBContainer = new MongoDBContainer(DockerImageName.parse("mongo:6.0.24"))
-		.withExposedPorts(MongoDB_PORT);
+	private static final PostgreSQLContainer<?> postgresContainer = new PostgreSQLContainer<>("postgres:14-alpine");
+
+	@DynamicPropertySource
+	static void registerProperties(DynamicPropertyRegistry registry) {
+		registry.add("spring.datasource.url", postgresContainer::getJdbcUrl);
+		registry.add("spring.datasource.username", postgresContainer::getUsername);
+		registry.add("spring.datasource.password", postgresContainer::getPassword);
+		registry.add("spring.datasource.driver-class-name", () -> "org.postgresql.Driver");
+	}
 
 	@Autowired
 	private ChatMemoryRepository chatMemoryRepository;
 
-	/**
-	 * Dynamically configure mongodb properties
-	 */
-	@DynamicPropertySource
-	static void registerProperties(DynamicPropertyRegistry registry) {
-		registry.add("spring.mongodb.host", mongoDBContainer::getHost);
-		registry.add("spring.mongodb.port", () -> mongoDBContainer.getMappedPort(MongoDB_PORT));
-	}
+	@Autowired
+	private JdbcTemplate jdbcTemplate;
 
 	@Test
 	void correctChatMemoryRepositoryInstance() {
-		assertThat(chatMemoryRepository).isInstanceOf(MongoDBChatMemoryRepository.class);
+		assertThat(chatMemoryRepository).isInstanceOf(ChatMemoryRepository.class);
 	}
 
 	@ParameterizedTest
@@ -81,12 +78,14 @@ class MongoDBChatMemoryRepositoryIT {
 
 		chatMemoryRepository.saveAll(conversationId, List.of(message));
 
-		var messages = chatMemoryRepository.findByConversationId(conversationId);
-		assertThat(messages).hasSize(1);
+		var query = "SELECT conversation_id, content, type, timestamp FROM ai_chat_memory WHERE conversation_id = ?";
+		var result = jdbcTemplate.queryForMap(query, conversationId);
 
-		var savedMessage = messages.get(0);
-		assertThat(savedMessage.getText()).isEqualTo(message.getText());
-		assertThat(savedMessage.getMessageType()).isEqualTo(messageType);
+		assertThat(result.size()).isEqualTo(4);
+		assertThat(result.get("conversation_id")).isEqualTo(conversationId);
+		assertThat(result.get("content")).isEqualTo(message.getText());
+		assertThat(result.get("type")).isEqualTo(messageType.name());
+		assertThat(result.get("timestamp")).isNotNull();
 	}
 
 	@Test
@@ -98,15 +97,20 @@ class MongoDBChatMemoryRepositoryIT {
 
 		chatMemoryRepository.saveAll(conversationId, messages);
 
-		var savedMessages = chatMemoryRepository.findByConversationId(conversationId);
-		assertThat(savedMessages.size()).isEqualTo(messages.size());
+		var query = "SELECT conversation_id, content, type, timestamp FROM ai_chat_memory WHERE conversation_id = ? ORDER BY timestamp";
+		var results = jdbcTemplate.queryForList(query, conversationId);
+
+		assertThat(results.size()).isEqualTo(messages.size());
 
 		for (var i = 0; i < messages.size(); i++) {
 			var message = messages.get(i);
-			var savedMessage = savedMessages.get(i);
+			var result = results.get(i);
 
-			assertThat(savedMessage.getText()).isEqualTo(message.getText());
-			assertThat(savedMessage.getMessageType()).isEqualTo(message.getMessageType());
+			assertThat(result.get("conversation_id")).isNotNull();
+			assertThat(result.get("conversation_id")).isEqualTo(conversationId);
+			assertThat(result.get("content")).isEqualTo(message.getText());
+			assertThat(result.get("type")).isEqualTo(message.getMessageType().name());
+			assertThat(result.get("timestamp")).isNotNull();
 		}
 
 		var count = chatMemoryRepository.findByConversationId(conversationId).size();
@@ -145,46 +149,39 @@ class MongoDBChatMemoryRepositoryIT {
 
 		chatMemoryRepository.deleteByConversationId(conversationId);
 
-		var results = chatMemoryRepository.findByConversationId(conversationId);
-		assertThat(results).isEmpty();
-	}
+		var count = jdbcTemplate.queryForObject("SELECT COUNT(*) FROM ai_chat_memory WHERE conversation_id = ?",
+				Integer.class, conversationId);
 
-	@Test
-	void clearOverLimit() {
-		var conversationId = UUID.randomUUID().toString();
-		var messages = List.<Message>of(new UserMessage("Message 1 from user - " + conversationId),
-				new AssistantMessage("Message 1 from assistant - " + conversationId),
-				new UserMessage("Message 2 from user - " + conversationId),
-				new AssistantMessage("Message 2 from assistant - " + conversationId),
-				new UserMessage("Message 3 from user - " + conversationId));
-
-		chatMemoryRepository.saveAll(conversationId, messages);
-
-		// Verify all messages have been saved
-		var savedMessages = chatMemoryRepository.findByConversationId(conversationId);
-		assertThat(savedMessages.size()).isEqualTo(messages.size());
-
-		// Perform cleanup operation, set max limit to 3, delete count to 2
-		MongoDBChatMemoryRepository mongoDBChatMemoryRepository = (MongoDBChatMemoryRepository) chatMemoryRepository;
-		mongoDBChatMemoryRepository.clearOverLimit(conversationId, 3, 2);
-
-		// Verify only the last 3 messages are retained
-		savedMessages = chatMemoryRepository.findByConversationId(conversationId);
-		assertThat(savedMessages.size()).isEqualTo(3);
-		assertThat(savedMessages.get(0).getText()).isEqualTo(messages.get(2).getText());
-		assertThat(savedMessages.get(1).getText()).isEqualTo(messages.get(3).getText());
-		assertThat(savedMessages.get(2).getText()).isEqualTo(messages.get(4).getText());
+		assertThat(count).isZero();
 	}
 
 	@SpringBootConfiguration
+	@ImportAutoConfiguration({ DataSourceAutoConfiguration.class, JdbcTemplateAutoConfiguration.class })
 	static class TestConfiguration {
 
+		@Autowired
+		private JdbcTemplate jdbcTemplate;
+
+		public void initializeDatabase() {
+			try {
+				jdbcTemplate.execute("DROP TABLE IF EXISTS ai_chat_memory");
+			}
+			catch (Exception e) {
+				// Ignore errors
+			}
+
+			jdbcTemplate.execute("CREATE TABLE ai_chat_memory (" + "id BIGSERIAL PRIMARY KEY, "
+					+ "conversation_id VARCHAR(256) NOT NULL, " + "content TEXT NOT NULL, "
+					+ "type VARCHAR(100) NOT NULL, " + "timestamp TIMESTAMP NOT NULL, "
+					+ "CONSTRAINT chk_message_type CHECK (type IN ('USER', 'ASSISTANT', 'SYSTEM', 'TOOL')))");
+
+			jdbcTemplate.execute("CREATE INDEX idx_conversation_id ON ai_chat_memory (conversation_id)");
+		}
+
 		@Bean
-		ChatMemoryRepository chatMemoryRepository() {
-			return MongoDBChatMemoryRepository.builder()
-				.host(mongoDBContainer.getHost())
-				.port(mongoDBContainer.getMappedPort(MongoDB_PORT))
-				.build();
+		ChatMemoryRepository chatMemoryRepository(JdbcTemplate jdbcTemplate) {
+			this.initializeDatabase();
+			return PostgresChatMemoryRepository.postgresBuilder().jdbcTemplate(jdbcTemplate).build();
 		}
 
 	}
